Document potential edge case with duplicate entity
**This Commit**
Makes note of the fact that we might clone+assign an `Aabb` go an entity
twice if it was assigned a new mesh and then re-assigned its old mesh.

**Why?**
Because the performance implications of checking for duplicates in the
rare case they exist outweigh the value of not cloning the `Aabb`.

See [this comment][0] for more details.

[0]: https://github.com/bevyengine/bevy/pull/4944#discussion_r890199315