Only register entities when Aabbs are calculated
**This Commit**
Updates `calculate_bounds` and `update_bounds` to only `register`
entity<->mesh relationships when an `Aabb` was successfully calculated
and assigned to the entity.

**Why?**
Mainly because if an entity has a mesh and we fail to compute its Aabb
in `calculate_bounds` then that entity will be present in the query the
next time `calculate_bounds` runs - we don't want to register it twice.
Basically we want `EntityMeshRelationships` to not just track "which
entities have which meshes" but instead "which entities with Aabbs have
which meshes".

**Note**
I've also removed the note about possible duplicates in the
`entities_with_mesh` map. I believe this is no longer possible since we
only `register` an entity with a mesh after deregistering it (or when it
gets an Aabb for the first time in which case it couldn't have been in
the list already).

I think this is correct - is there an edge case I'm missing?